move : 서치 디렉토리 변경
