+++ src/features/board/ui/search/searchBar/SearchBar.style.ts
