+++ src/features/board/ui/search/stackSearch/StackSearch.style.ts
